moved settings pannel to components dir
